+++ core/math/math_fieldwise.cpp
